# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/stores/{store_id}/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products for a specific store",
                "operationId": "aa4f34a732143f2379babab37e868fe2",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Store not found"
                    }
                }
            }
        },
...
        "/stores": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get all stores",
                "description": "Mengambil semua data toko",
                "operationId": "getStores",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data toko",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Store"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Create a new store",
                "description": "Membuat toko baru",
                "operationId": "createStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toko Elektronik"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "example": "http://example.com/image.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Store berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Store"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request tidak valid"
                    }
                }
            }
        },
        "/stores/{id}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get a store by ID",
                "description": "Mengambil data toko berdasarkan ID",
                "operationId": "getStoreById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID toko",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Toko ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Store"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Toko tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stores"
                ],
                "summary": "Update store data",
                "description": "Memperbarui data toko berdasarkan ID",
                "operationId": "updateStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID toko",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toko Gadget"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "example": "http://example.com/gadget.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Toko berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Store"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Toko tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store",
                "description": "Menghapus toko berdasarkan ID",
                "operationId": "deleteStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID toko",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Toko berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Store deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Toko tidak ditemukan"
                    }
                }
            }
        },
        "/stores/{id}/products": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get store with its products",
                "description": "Mengambil data toko beserta produk-produknya",
                "operationId": "getStoreWithProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID toko",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data toko beserta produknya",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Store"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Toko tidak ditemukan"
                    }
                }
            }
        },
        "/stores/search": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Search stores by name",
                "description": "Mencari toko berdasarkan nama",
                "operationId": "searchStoresByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nama toko yang ingin dicari",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Elektronik"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman yang ingin ditampilkan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mencari toko",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stores/city/{city}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get stores by city",
                "description": "Mengambil data toko berdasarkan kota",
                "operationId": "getStoresByCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "Nama kota",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bandung"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman yang ingin ditampilkan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data toko berdasarkan kota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product data schema",
                "required": [
                    "product_id",
                    "name",
                    "price",
                    "stock",
                    "status",
                    "store_id",
                    "category_id"
                ],
                "properties": {
                    "product_id": {
                        "description": "Product primary key",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Smartphone X200"
                    },
                    "thumbnail_image": {
                        "description": "Path to the thumbnail image",
                        "type": "string",
                        "example": "images/products/thumb.jpg",
                        "nullable": true
                    },
                    "stock": {
                        "description": "Current stock quantity",
                        "type": "integer",
                        "example": 50
                    },
                    "status": {
                        "description": "Availability status of the product",
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "description": {
                        "description": "Detailed product description",
                        "type": "string",
                        "example": "Latest smartphone with high-resolution display",
                        "nullable": true
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "format": "float",
                        "example": 499.99
                    },
                    "store_id": {
                        "description": "Foreign key referencing store",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "Foreign key referencing category",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "description": "Record creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Record last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-24T08:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "Store": {
                "title": "Store",
                "required": [
                    "name",
                    "city"
                ],
                "properties": {
                    "store_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Toko Elektronik"
                    },
                    "city": {
                        "type": "string",
                        "example": "Bandung"
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-30T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-30T12:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}